QXmlStreamReader: fix integer truncation (qint64 → int)
We don't care about the actual type of oldLineNumber here, we just
need it to be of the same type as lineNumber, to avoid truncation, so
we use auto.

This was already broken in 5.15 (lineNumber was already a qint64).

Change-Id: I68d7e6e91c8122bf426c9c10e8cc91ff55c61a20
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 389fdd6d8d1f51c026b82a3cb4e2379c3d9debf1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>