QXmlStreamReader: port Value::pos from int to qsizetype
It's an index into a QString, so it has to be qsizetype.

The length and prefix fields also shouldn't stay ints, but that's a
different patch, because they're just relative offsets to pos, and so
aren't as easily overflown.

This enlarges the Value struct; created QTBUG-103306 to track ideas to
shrink it again.

Pick-to: 6.3 6.2
Task-number: QTBUG-102465
Change-Id: Ib318e131420c2c535e26cb03cbf450e031626e64
Reviewed-by: Lars Knoll <lars.knoll@qt.io>