Update commercial license headers
Updated header.COMM to the files in tqtc-qtbase.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4934
Change-Id: Ie60b1b5f64c28ddcc77462abadbcf66af7c6d73b
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>